tools/b64_encoder/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]

[tool.setuptools_scm]
root = "../.."

[project]
name = "b64_encoder"
dynamic = ["version"]
dependencies = [
  "nvidia-nat[langchain]",
]
requires-python = ">=3.11,<3.13"
description = "Custom NeMo Agent Toolkit Workflow"
classifiers = ["Programming Language :: Python"]

[tool.uv.sources]
nvidia-nat = { path = "../..", editable = true }

[project.entry-points.'nat.components']
b64_encoder = "b64_encoder.register"